Update pattern editor

+++ app/src/main/java/libre/sampler/models/ScheduledNoteEvent.java
@@ -8,9 +8,9 @@ public class ScheduledNoteEvent {
     private int action;
     private int keyNum;
     private int velocity;
-    private int baseId;
+    private long baseId;
 
-    public ScheduledNoteEvent(Long offsetTicks, int action, int keyNum, int velocity, int baseId) {
+    public ScheduledNoteEvent(Long offsetTicks, int action, int keyNum, int velocity, long baseId) {
         this.offsetTicks = offsetTicks;
         this.action = action;
         this.keyNum = keyNum;
@@ -19,6 +19,6 @@ public class ScheduledNoteEvent {
     }
 
     public NoteEvent getNoteEvent(int loopIndex) {
-        return new NoteEvent(action, keyNum, velocity, new Pair<>((long) baseId, loopIndex));
+        return new NoteEvent(action, keyNum, velocity, new Pair<>(baseId, loopIndex));
     }
 }
